% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aRD-class.R
\docType{class}
\name{aRD-class}
\alias{aRD-class}
\title{S4 Class \code{"aRD"}}
\description{
S4 Class \code{"aRD"}
}
\section{Slots}{

\describe{
\item{\code{call}}{An object of class \code{"call"}.}

\item{\code{formula}}{An object of class \code{"formula"}.}

\item{\code{coefficients}}{A numeric vector containing the estimated model parameters.}

\item{\code{iter}}{A positive integer indicating the number of iterations.}

\item{\code{converged}}{A logical constant that indicates whether the model has converged.}

\item{\code{y}}{A numeric vector containing the dependent variable of the model.}

\item{\code{x}}{The model matrix.}

\item{\code{data}}{A data frame containing the variables in the model.}
}}

\author{
Thomas Wolf, Julius Johannes Weise, Stefan Wagenpfeil
}
