% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARD.R
\name{ARD}
\alias{ARD}
\title{Automatic Relevance Determination}
\usage{
ARD(X, Y, sigma.square = estimateVariance(X, Y),
  sigma.square.init = as.numeric(var(Y)), gamma.init = rep(1, ncol(X)),
  eta.init = rep(1, ncol(X)), threshold_eta = 10000, maxstep = 1000,
  margin = 0.001, verbosity = 1, compute.likelihood = TRUE,
  method = "FixedPoint")
}
\arguments{
\item{X}{predictor matrix of dimension \eqn{n*q}.}

\item{Y}{continuous outcome vector of dimension \eqn{p}.}

\item{sigma.square}{variance of noise, used for method 'Reweighted l1'. Default value is estimated by \code{estimateVariance(X,Y)}}

\item{sigma.square.init}{variance of noise, used for method 'FixedPoint' and 'EM'. Default value is \code{var(Y)}
initial_val initial value for \eqn{\alpha}. Default value is a vector of \eqn{0}s.}

\item{gamma.init}{initial value for hyper parameter, only for method 'Reweighted l1'}

\item{eta.init}{initial value for hyper parameter, only for method 'EM' and 'FixedPoint'}

\item{threshold_eta}{cut off point for very large eta}

\item{maxstep}{max number of iterations}

\item{margin}{iteration stoping creteria}

\item{verbosity}{print current iteraction number}

\item{compute.likelihood}{whether compute likelihood in each step}
}
\value{
returns the estimated coefficients, estimated hyper-parameters, number of iterations and likelihood scores by Automated Relevance Determination (ARD).
}
\description{
\code{ARD} returns the estimated coefficients, estimated hyper-parameters, number of iterations and likelihood scores by Automated Relevance Determination (ARD).
}
\details{
ARD based on three optimization algorithms -- EM algorithm, Fixed Point update rule and Reweighted-l1 algorithm.
If use method 'reweighted-l1', please provide sigma.square otherwise an external function was called to estimate sigma square
if you have prefered initial value for gamma, set gamma.init, otherwise gamma.init was initialized to be rep(1,ncol(X))
If use method: EM or Fixed Point update,sigma.square.init was used,the default initial value is variance of Y,if you have prefered initial value for eta, set eta.init, otherwise eta.init was initialized to be rep(1,ncol(X))
}
\examples{
set.seed(99)
n = 100
p = 200
x <- matrix(rnorm(2*n*p,0,1),nrow=2*n,ncol=p)
betas=rnorm(n = p, s = 1/exp(z_design\%*\%alpha))
y <- x\%*\%betas + rnorm(2*n,0,1)
x_train = x[1:n,]
y_train = y[1:n]
x_test = x[(n+1):(2*n),]
y_test = y[(n+1):(2*n)]
ARD.fit = ARD(x_train,y_train)
1-get_mse(x_test\%*\%ARD_fit$coef,y_test)/var(y_test)
}
