% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef-method.R
\name{coef,aRD-method}
\alias{coef,aRD-method}
\title{Extracting the estimated model parameters of \code{aRD()}}
\usage{
\S4method{coef}{aRD}(object)
}
\arguments{
\item{object}{An object of class \code{"aRD"}.}
}
\value{
A numeric vector containing the estimated model parameters.
}
\description{
For objects of class \code{"aRD"}, \code{coef()} extracts the estimated model parameters of \code{aRD()}.
}
\author{
Thomas Wolf, Julius Johannes Weise, Stefan Wagenpfeil
}
