% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constr.R
\name{constr}
\alias{constr}
\title{Setting the linear inequality constraints for \code{aRD()}}
\usage{
constr(x, version = 1)
}
\arguments{
\item{x}{A model matrix.}

\item{version}{switch for constraints}
}
\value{
A matrix containing the linear inequality constraints for \code{aRD()}.
}
\description{
\code{constr()} sets the linear inequality constraints for \code{aRD()}.
}
\author{
Thomas Wolf, Julius Johannes Weise, Stefan Wagenpfeil
}
