% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradF.R
\name{gradF}
\alias{gradF}
\title{Deriving the first derivatives of the log likelihood function of the linear-binomial model in \code{aRD()}}
\usage{
gradF(theta, y, x)
}
\arguments{
\item{theta}{A numeric vector containing the initial values of the model parameters.}

\item{y}{A numeric vector containing the dependent variable of the model.}

\item{x}{The model matrix.}
}
\value{
A numeric vector containing the first derivatives of the log likelihood function of the linear-binomial model.
}
\description{
\code{gradF()} derives the first derivatives of the log likelihood function of the linear-binomial model.
}
\author{
Thomas Wolf, Julius Johannes Weise, Stefan Wagenpfeil
}
