% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hess.R
\name{hess}
\alias{hess}
\title{Deriving the second partial derivatives of the log likelihood function of the linear-binomial model in \code{aRD()} (Hessian matrix)}
\usage{
hess(theta, y, x)
}
\arguments{
\item{theta}{A numeric vector containing the initial values of the model parameters.}

\item{y}{A numeric vector containing the dependent variable of the model.}

\item{x}{The model matrix.}
}
\value{
A numeric matrix containing the second partial derivatives of the log likelihood function of the linear-binomial model (Hessian matrix).
}
\description{
\code{aRD()} derives the second partial derivatives of the log likelihood function of the linear-binomial model.
}
\author{
Thomas Wolf, Julius Johannes Weise, Stefan Wagenpfeil
}
