% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_value.R
\name{obj_value}
\alias{obj_value}
\title{Log-Likelihood Function Value for the linear-binomial model in \code{aRD()}}
\usage{
obj_value(theta, y, x)
}
\arguments{
\item{theta}{A numeric vector containing the initial values of the model parameters.}

\item{y}{A numeric vector containing the dependent variable of the model.}

\item{x}{The model matrix.}
}
\value{
Numeric scalar representing the value of the log-likelihood function evaluated at \code{theta}.
}
\description{
\code{obj_value()} computes the value of the log-likelihood function for the linear-binomial model
at a given parameter vector \code{theta}.
}
\author{
Thomas Wolf, Julius Johannes Weise, Stefan Wagenpfeil
}
