% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARD plot.R
\name{plot.ARD}
\alias{plot.ARD}
\title{Automatic Relevance Determination Plots}
\usage{
\method{plot}{ARD}(ARD.result)
}
\arguments{
\item{ARD.result}{ARD object}
}
\value{
marginal likelihood trace plot and the coefficient plot likelihood scores by Automated Relevance Determination (ARD).
}
\description{
\code{plot.ARD} returns two plots -- the marginal likelihood trace plot and the coefficient plot
}
\examples{
plot.ARD(ARD.result)
}
