% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARD_predict.R
\name{predict.ARD}
\alias{predict.ARD}
\title{Automatic Relevance Determination Prediction}
\usage{
\method{predict}{ARD}(ARD.result, data.to.predict)
}
\arguments{
\item{ARD.result}{ARD object}

\item{data.to.predict}{prediction data}
}
\value{
predicted value
}
\description{
\code{predict.ARD} returns predicted value
}
\examples{
predict.ARD(ARD.result,x_test)
}
