% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-method.R
\name{summary,aRD-method}
\alias{summary,aRD-method}
\title{Summarizing the estimated model parameters of \code{aRD()}}
\usage{
\S4method{summary}{aRD}(object)
}
\arguments{
\item{object}{An object of class \code{"aRD"}.}
}
\value{
A list containing the following elements:
\item{coefficients}{A numeric vector containing the estimated model parameters.}
\item{std.err}{A numeric vector containing the estimated standard errors of the model parameters.}
\item{z.value}{A numeric vector containing the estimated z test statistic of the model parameters.}
\item{p.value}{A numeric vector containing the estimated p values of the model parameters.}
}
\description{
For objects of class \code{"aRD"}, \code{summary()} summarizes the estimated model parameters of \code{aRD()}.
}
\author{
Thomas Wolf, Julius Johannes Weise, Stefan Wagenpfeil
}
